@@ -3,17 +3,17 @@
   "version": "0.1.0",
   "private": true,
   "scripts": {
-    "dev": "next dev",
-    "build": "next build",
-    "start": "next start",
+    "dev": "next dev --webpack",
+    "build": "next build --webpack",
+    "start": "next start --webpack",
     "check:format": "prettier --check src",
-    "check:lint": "next lint",
+    "check:lint": "eslint .",
     "check:types": "tsc --pretty --noEmit",
     "check:styles": "stylelint 'src/**/*.tsx'",
     "clean:modules": "rm -rf node_modules",
     "clean": "rm -rf .next && rm -rf node_modules && pnpm i",
     "format": "prettier --write .",
-    "lint:fix": "next lint --fix",
+    "lint:fix": "eslint --fix .",
     "styles:fix": "stylelint 'src/**/*.tsx' --fix",
     "test:all": "pnpm run check:format && pnpm run check:styles && pnpm run check:lint && pnpm run check:types && pnpm run build",
     "prepare": "husky || true"
@@ -24,8 +24,8 @@
     "@linaria/react": "^6.3.0",
     "@mdx-js/loader": "^3.1.1",
     "@mdx-js/react": "^3.1.1",
-    "@next/bundle-analyzer": "^15.5.3",
-    "@next/mdx": "^15.5.3",
+    "@next/bundle-analyzer": "16.0.10",
+    "@next/mdx": "16.0.10",
     "@t3-oss/env-nextjs": "^0.12.0",
     "@tanstack/react-query": "^5.89.0",
     "@tanstack/react-query-devtools": "^5.89.0",
@@ -40,48 +40,66 @@
     "gray-matter": "^4.0.3",
     "lodash.filter": "^4.6.0",
     "milliseconds": "^1.0.3",
-    "motion": "^12.23.16",
-    "next": "^15.5.3",
+    "motion": "^12.23.15",
+    "next": "16.0.10",
     "next-compose-plugins": "^2.2.1",
     "next-mdx-remote": "^5.0.0",
     "next-with-linaria": "^1.2.0",
     "open-props": "^1.7.16",
     "openmeteo": "^1.2.0",
     "pokenode-ts": "^1.20.0",
-    "react": "^19.1.1",
-    "react-dom": "^19.1.1",
+    "react": "19.2.3",
+    "react-dom": "19.2.3",
     "react-icons": "^5.5.0",
     "rehype-prism-plus": "^2.0.1",
     "remark-smartypants": "^3.0.2",
     "zod": "^4.1.10"
   },
   "devDependencies": {
+    "@eslint/compat": "2.0.0",
+    "@eslint/js": "9.39.2",
     "@linaria/babel-preset": "^5.0.4",
     "@linaria/postcss-linaria": "^6.3.0",
     "@linaria/stylelint": "^6.3.0",
     "@linaria/webpack-loader": "^5.0.4",
+    "@next/eslint-plugin-next": "16.0.10",
+    "@stylistic/eslint-plugin": "5.6.1",
     "@tanstack/eslint-plugin-query": "^5.89.0",
+    "@types/eslint": "9.6.1",
     "@types/lodash.filter": "^4.6.9",
     "@types/mdx": "^2.0.13",
     "@types/milliseconds": "^1.0.0",
     "@types/node": "^22.18.6",
-    "@types/react": "^19.1.13",
-    "@types/react-dom": "^19.1.9",
-    "eslint": "^9.36.0",
-    "eslint-config-next": "15.3.3",
+    "@types/react": "19.2.7",
+    "@types/react-dom": "19.2.3",
+    "eslint": "^9.39.2",
+    "eslint-config-next": "16.0.10",
     "eslint-config-prettier": "^10.1.8",
+    "eslint-plugin-jsx-a11y": "6.10.2",
+    "eslint-plugin-mdx": "3.6.2",
+    "eslint-plugin-react": "7.37.5",
+    "eslint-plugin-react-compiler": "19.1.0-rc.2",
+    "eslint-plugin-react-hooks": "7.0.1",
     "husky": "^9.1.7",
+    "jiti": "2.6.1",
     "lint-staged": "^16.1.6",
     "postcss-jsx": "^0.36.4",
     "postcss-styled-syntax": "^0.7.1",
     "prettier": "^3.6.2",
     "stylelint": "^16.24.0",
     "stylelint-config-recommended": "^16.0.0",
-    "typescript": "^5.9.2"
+    "typescript": "^5.9.2",
+    "typescript-eslint": "8.50.0"
   },
   "lint-staged": {
     "*.js": "eslint --cache --fix",
     "*.{js,css,md}": "prettier --write"
   },
-  "packageManager": "pnpm@10.11.1"
+  "packageManager": "pnpm@10.11.1",
+  "pnpm": {
+    "overrides": {
+      "@types/react": "19.2.7",
+      "@types/react-dom": "19.2.3"
+    }
+  }
 }
