@@ -1,9 +1,9 @@
 import { css } from '@linaria/core';
 
-type CardProps = {
+interface CardProps {
   title: string;
   description: string;
-};
+}
 
 export default function Card({ title, description }: CardProps) {
   return (
