@@ -6,7 +6,7 @@ import FeatureCollection from '@/types/GeoapifyResponse';
 
 const fetchGeolocation = async (lat: number, lng: number) => {
   // https://apidocs.geoapify.com/docs/geocoding/reverse-geocoding/
-  const response = axios.get(`https://api.geoapify.com/v1/geocode/reverse?lat=${lat}&lon=${lng}&apiKey=${env.NEXT_PUBLIC_GEOAPIFY_API_KEY}`);
+  const response = axios.get(`https://api.geoapify.com/v1/geocode/reverse?lat=${lat.toFixed()}&lon=${lng.toFixed()}&apiKey=${env.NEXT_PUBLIC_GEOAPIFY_API_KEY}`);
   const data = (await response).data as FeatureCollection;
 
   return data;
